refactor some tests
