+++ tests/pw/e2e.config.ts
@@ -9,7 +9,7 @@ export default defineConfig({
     // globalTeardown: './global-teardown' /* Path to the global teardown file. This file will be required and run after all the tests. */,
     globalTimeout: process.env.CI ? 40 * (60 * 1000) : 40 * (60 * 1000) /* Maximum time in milliseconds the whole test suite can run */,
     maxFailures: process.env.CI ? 40 : 30 /* The maximum number of test failures for the whole test suite run. After reaching this number, testing will stop and exit with an error. */,
-    timeout: process.env.CI ? 50 * 1000 : 35 * 1000 /* Maximum time one test can run for. */,
+    timeout: process.env.CI ? 60 * 1000 : 35 * 1000 /* Maximum time one test can run for. */,
     expect: {
         timeout: 15 * 1000 /* Maximum time expect() should wait for the condition to be met.  For example in `await expect(locator).toHaveText();`*/,
         toHaveScreenshot: {
